recommendations/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
numpy
